Bug fix in set _fake, And varies utilities for fixing pipeline
major blunder in the set_fake function,root._fields was not set!

At the same time,mark_under_fixing  and under_fixing has been added In order to Annotate nodes That are under fixing. Comes in useful, when fixing a node Requires fixing other nodes, To avoid an infinite loop