Bug fix. add check To prevent off by one errorsThese are the boundaries of functions
While normally, if on an empty line origin is going to bound to the line below, On the boundaries between functions  It will bound upwards by design.

Unfortunately, These caused An issue because to prevent all my one errors When going upwards When the statements spread over multiple lines, there were checks  if we were below its beginning but not thief we were above its ending( because otherwise Origin would normally bound to the linee below!).

These commit Add those checks and resolve #28