add invert_then_tiebreak argument to process_line Argument Queries
This addition was neededSo the queries of the form "inside nth argument" would work properly. In particular, for these type of queries tiebreaking Needs to happen after inversion because we need to pick the closest call that satisfies the nth description with respect to its (local) parent. Otherwise, the result produced would not be even deterministic,as helpful will not be ordered!!!